Update lmo-tippopenfileall.php
PHP Warning:  A non-numeric value encountered in ../lmo/addon/tipp/lmo-tippopenfileall.php on line 78
Bei eingeloggtem User und Aufruf der Tipp-Tabelle

+++ lmo/addon/tipp/lmo-tippopenfileall.php
@@ -75,7 +75,7 @@ for($i = 1; $i <= count($tippdaten); $i++) {
   $op2 = substr($dum[0], 0, 5);
   $op3 = substr($dum[0], 5)-1;
   $op4 = substr($dum[1], 2)-1;
-  $op6 = substr($dum[1], 2, -1)-1;
+  $op6 = (int)substr($dum[1], 2, -1)-1;
   $op7 = substr($dum[1], -1)-1;
   $op8 = substr($dum[1], 0, 2);
   $jksp[$op3] = $dum[3];
